Added Tabletop class
An "arena" is injected as a dependency via the Robot.place method.

Arena's must respont to #move_piece(x,y,facing,1) . The last argument is
optional and denotes the number of points to travel forward or backwards (e.g.
1, -1).

NOTE: Both classes must be aware of the points of a compass (North, South,
East, West).